Data Science Projects/VisionScout/Deployment/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.7.0
pillow>=9.4.0
numpy>=1.23.5
matplotlib>=3.7.0
gradio>=3.32.0
git+https://github.com/openai/CLIP.git
yt-dlp>=2023.3.4
requests>=2.28.1
transformers
accelerate
bitsandbytes
sentencepiece
huggingface_hub>=0.19.0
urllib3>=1.26.0
